Rename `validation` module
