+++ crates/smart-config/src/metadata/_private.rs
